requirements.txt
requests
gitpython
python-dotenv
